[REG-1363] Update generated scripts to new conventions
