--- Assets/RGScripts/BotCharacterConfig.cs
@@ -1,8 +1,0 @@
-
-using System;
-
-[Serializable]
-public class BotCharacterConfig
-{
-    public float speed;
-}
