--- Assets/RGScripts/BotCharacterConfig.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: c27abbf9ff5149a0b74721f127254d92
-timeCreated: 1687997839
