--- Assets/RGScripts/RGSerialization.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2aecd365a92a6435cb274fe97e59ea92
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
