+++ Assets/RegressionGames/Runtime.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 941af9ae6369f4b639a02dadfcc1c92a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
