+++ Assets/RegressionGames/Runtime/GeneratedScripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d42d861279a004f8fbf9a21256e408cc
+guid: e12dd4c0e686d4f39977e60193aad2aa
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
