+++ Assets/RegressionGames/Runtime/GeneratedScripts/RGSerialization.cs
@@ -3,6 +3,7 @@
 */
 
 using System;
+using Newtonsoft.Json;
 
 namespace RegressionGames
 {
