+++ Assets/RegressionGames/Runtime/GeneratedScripts/RGSerialization.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 450ed1f82e8684d038d97edb224e5850
+guid: 529b130f65b8e4f1ea617451a2e5d16c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
