+++ Assets/RegressionGames/Runtime/GeneratedScripts/RGStates.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8877a438b7e574750ba49de0352ee107
+guid: 62e15a821cd1c4bf0bcc21722afc838a
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
