+++ Assets/RegressionGames/Runtime/GeneratedScripts/RGStates/RGState_Player.cs
@@ -4,10 +4,20 @@
 
 using System;
 using System.Collections.Generic;
+using RegressionGames;
+using RegressionGames.RGBotConfigs;
+using RegressionGames.StateActionTypes;
+using UnityEngine;
 
 namespace RegressionGames.RGBotConfigs
 {
-    public class Player_RGState : RGState
+    public class RGStateEntity_Player : RGStateEntity<RGState_Player>
+    {
+        public float speed => (float)float.Parse(this.GetValueOrDefault("speed").ToString());
+        public float range => (float)float.Parse(this.GetValueOrDefault("range").ToString());
+    }
+
+    public class RGState_Player : RGState
     {
         private Player myComponent;
         public void Start()
