+++ Assets/RegressionGames/Runtime/GeneratedScripts/RGStates/RGState_Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2d59cb9a4c5484406ae4b1f2da6bcaed
+guid: 53cd9fed1051345a6ab61c91d7551d22
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
